Gas Assist
A new system that makes the player move more in the direction of the grapple instead of around it. This will give more control, but will give you less speed while swinging. I recommend keeping gasAssistForce at most round 10 . When set to 100 this will cause the player to move in almost straight lines straight to the grapple.